# server/chat-it/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Chatlt
# during dev, should point to your local machine
host: 50.112.200.45:80
# basePath prefixes all resource paths
basePath: /
#
schemes:
  # tip: remove http to make production-grade
  - http
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /webhook:
    x-swagger-router-controller: chat-it
    post:
      description: webhook for api.aijquery
      operationId: webhook
      parameters:
      - name: webhook
        in: body
        description: webhook body response
        schema:
          properties:
            webhook:
              type: object
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/WebhookResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
            
            
  /postItinerary:
    x-swagger-router-controller: chat-it
    post:
      description: Post itinerary form data
      # used as the method name of the controller
      operationId: postItinerary
      parameters:
        - name: itineraryData
          in: body
          description: Post itinerary form data, postItinerary.data is an array of days, with each day containing an array of hours of that day. Within each array of hours there is a JSON object containing what the user did. Ex. location, activity, time of activity, duration of activity.
          schema:
            $ref: "#/definitions/postItinerary"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /getItinerary:
    x-swagger-router-controller: chat-it
    get:
      description: get result of itinerary based on filters, returns 10 results sorted by popularity if nothing given
      operationId: getItinerary
      parameters:
      - name: getParams
        in: body
        description: Takes a query object and returns itinerary results
        schema:
          properties:
            query:
              type: object
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /login:
    x-swagger-router-controller: chat-it
    post:
      description: Post login
      # used as the method name of the controller
      operationId: login
      parameters:
      - name: loginData
        in: body
        description: login data
        schema:
          $ref: "#/definitions/loginData"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /register:
    x-swagger-router-controller: chat-it
    post:
      description: new user registration
      # used as the method name of the controller
      operationId: register
      parameters:
      - name: registerData
        in: body
        description: registration data
        schema:
          $ref: "#/definitions/registerData"

      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /review:
    x-swagger-router-controller: chat-it
    post:
      description: user posting review for the itinerary
      # used as the method name of the controller
      operationId: postReview
      parameters:
        - name: itineraryID
          in: query
          description: ID of the itinerary to fetch
          required: false
          type: string
        - name: title
          in: query
          description: title of the comment
          required: true
          type: string
        - name: data
          in: body
          description: review
          schema:
            $ref: "#/definitions/postReview"

      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"



  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:

  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string

  SuccessResponse:
    required:
      - message
    properties:
      message:
        type: string
  entry:
    required:
      - location
      - activity
      - time
      - duration
    properties:
      location:
        type: string
      activity:
        type: string
      time:
        type: string
      duration:
        type: integer

  postItinerary:
    required:
      - location
      - dateFrom
      - dateTo
      - itineraryData
    properties:
      location:
        type: string
      dateFrom:
        type: string
      dateTo:
        type: string
      itineraryData:
        type: array
        items:
          type: array
          items:
            $ref: "#/definitions/entry"
  registerData:
      required:
        - username
        - password
        - email
        - phoneNumber
      properties:
        username:
          type: string
        password:
          type: string
        email:
          type: string
        phoneNumber:
         type: number
  loginData:
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string

  WebhookResponse:
    required:
      - speech
      - displayText
      - data
      - contextOut
      - source
    properties:
      speech:
        type: string
      displayText:
        type: string
      data:
        type: object
      contextOut:
        type: array
        items:
          properties:
            location:
              type: string
            date:
              type: string
      source:
        type: string
        
  postReview:
    required:
      - data
    properties:
      data:
        type: string
        items:
           $ref: "#/definitions/postReview"